Add and update new classes and functionality
Created new classes such as ToDo, Deadline and Event to
make known the different types of tasks.

A superclass Task is inherited by the child classes to
allow inheritance of pre-existing instances to make it
more neat and easy to comprehand.

To carry out a Text Ui Testing followed by reviewing code
to aline with code quality standards.