Add validation to markOrUnmarkTask
A new class have been added called DukeException. This class is used
to track for exceptions that are specific to Duke.java program.

Streamline the validation for typeOfTask, in coporated extra validation
to check and if it user input spaces instead of leaving it blank.

To address the user input for taskDetails next and validate to see if
user have typed in the event detail correctly or have left blank.
If time permit, else leave validation for wk6 ip