Add Increment Level-5 Class IllegalInputException to handle illegal inputs
Users are able to input illegal inputs and commands that will crash the program or make it behave not as expected.

An IllegalInputException should be thrown whenever the user enters an illegal input because users should not be able to crash the program or cause bugs in the program.

Let's update the methods parsing user inputs to throw errors and the method handling user input to catch IllegalInputExceptions and display them on the terminal to the user.