Add error handling for delete
Added error handling when user does not specify what to delete or
specify something that is unknown.

Updated the print list to include a message that notify the user if
there are no task in the list at present.

Will create tag for Level-6 and A-Collections and will create new
breanch-Level-7 in master for Level-7 task.